Move checkValidations to shared
