+++ src/applications/appeals/995/tests/validations/validations.unit.spec.js
@@ -1,20 +1,8 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { checkValidations, missingPrimaryPhone } from '../../validations';
-import { missingIssueName } from '../../validations/issues';
+import { missingPrimaryPhone } from '../../validations';
 import { errorMessages, PRIMARY_PHONE } from '../../constants';
-
-describe('checkValidations', () => {
-  it('should return error messages', () => {
-    expect(checkValidations([missingIssueName], '')).to.deep.equal([
-      errorMessages.missingIssue, // simple validation function
-    ]);
-    expect(
-      checkValidations([missingIssueName, missingIssueName], ''),
-    ).to.deep.equal([errorMessages.missingIssue, errorMessages.missingIssue]);
-  });
-});
 
 describe('missingPrimaryPhone', () => {
   it('should show an error if no primary phone selected', () => {
