+++ src/applications/appeals/shared/validations/issues.js
@@ -1,21 +1,6 @@
 import { hasDuplicates, hasSomeSelected } from '../utils/issues';
 import { noneSelected } from '../content/contestableIssues';
 import errorMessages from '../content/errorMessages';
-/**
- *
- * @param {Function[]} validations - array of validation functions
- * @param {*} data - field data passed to the validation function
- * @param {*} fullData - full and appStateData passed to validation function
- * @returns {String[]} - error messages
- */
-export const checkValidations = (validations, data, fullData) => {
-  const errors = { errorMessages: [] };
-  errors.addError = message => errors.errorMessages.push(message);
-  validations.map(validation =>
-    validation(errors, data, fullData, null, null, null, fullData),
-  );
-  return errors.errorMessages;
-};
 
 export const selectionRequired = (
   errors,
@@ -47,5 +32,11 @@ export const uniqueIssue = (
 ) => {
   if (errors?.addError && hasDuplicates(appStateData || formData)) {
     errors.addError(errorMessages.uniqueIssue);
+  }
+};
+
+export const missingIssueName = (errors, data) => {
+  if (!data) {
+    errors.addError(errorMessages.missingIssue);
   }
 };
